Going over the code. Using Xcode 5. Bumping the minimum OS X version to 10.8.

+++ Cocoa/Plug-Ins/Template Editor/ElementDLNG.m
@@ -41,7 +41,7 @@
 
 - (NSString *)stringValue
 {
-	return [NSString stringWithFormat:@"%d", value];
+	return [NSString stringWithFormat:@"%d", (int)value];
 }
 
 - (void)setStringValue:(NSString *)str
